requirements.txt
web3
easysolc
easyweb3
python-dotenv
